requirements.txt
# Core dependencies
openai>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
tiktoken>=0.5.0

# Visualization
matplotlib>=3.7.0

# Optional dependencies for enhanced functionality
python-dotenv>=1.0.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0